+++ src/main/java/wooteco/subway/exception/NoSuchMemberException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.web.member;
+package wooteco.subway.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
