+++ src/main/java/wooteco/subway/service/member/MemberService.java
@@ -9,15 +9,14 @@ import wooteco.subway.domain.station.Station;
 import wooteco.subway.domain.station.StationRepository;
 import wooteco.subway.infra.JwtTokenProvider;
 import wooteco.subway.service.favorite.dto.FavoriteCreateRequest;
-import wooteco.subway.service.favorite.dto.FavoriteReadRequest;
 import wooteco.subway.service.favorite.dto.FavoriteResponse;
 import wooteco.subway.service.member.dto.LoginRequest;
 import wooteco.subway.service.member.dto.UpdateMemberRequest;
 import wooteco.subway.service.station.NoSuchStationException;
-import wooteco.subway.web.member.DuplicateMemberException;
-import wooteco.subway.web.member.InvalidAuthenticationException;
-import wooteco.subway.web.member.NoSuchFavoriteException;
-import wooteco.subway.web.member.NoSuchMemberException;
+import wooteco.subway.exception.DuplicateMemberException;
+import wooteco.subway.exception.InvalidAuthenticationException;
+import wooteco.subway.exception.NoSuchFavoriteException;
+import wooteco.subway.exception.NoSuchMemberException;
 
 import java.util.ArrayList;
 import java.util.List;
