+++ src/test/java/wooteco/subway/web/member/login/BearerAuthInterceptorTest.java
@@ -2,14 +2,13 @@ package wooteco.subway.web.member.login;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 import wooteco.subway.infra.JwtTokenProvider;
-import wooteco.subway.web.member.InvalidAuthenticationException;
+import wooteco.subway.exception.InvalidAuthenticationException;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
